✨ Read notification on click, make grade double, redesign admin edit user empty state

+++ backend/UBBGradePortal.Application/DTOs/SolvedActivity/GradeSolvedActivityDto.cs
@@ -6,5 +6,5 @@ namespace UBBGradePortal.Application.DTOs.SolvedActivity;
 public record GradeSolvedActivityDto(
     SolvedActivityStatus Status,
     string? ProfessorComment,
-    int Grade
+    double Grade
 );
